fix: use AtomicI64 for strategy_sums to prevent overflow
Strategy sums accumulate monotonically (especially after LCFR warmup
when discounting stops) and overflow i32 within a few million iterations,
causing average strategies to collapse to uniform.

Changes:
- storage.rs: strategy_sums Vec<AtomicI32> -> Vec<AtomicI64>
- optimizer.rs: apply_discount takes &[AtomicI64] for strategy_sums
- mccfr.rs: cast strategy sum delta as i64
- trainer.rs: prev_strategy_sums Vec<i32> -> Vec<i64>
- Serialization saves (i32 regrets, i64 sums), loads legacy formats

Regrets stay as AtomicI32 -- they get discounted and floored, so they
don't overflow.

+++ crates/core/src/blueprint_v2/mccfr.rs
@@ -965,7 +965,7 @@ fn traverse_traverser(
 
     // Accumulate strategy sums (for computing the average strategy).
     for (a, &s) in strategy.iter().enumerate().take(num_actions) {
-        storage.add_strategy_sum(node_idx, bucket, a, (s * 1000.0) as i32);
+        storage.add_strategy_sum(node_idx, bucket, a, (s * 1000.0) as i64);
     }
 
     // Accumulate EV at tracked scenario nodes.
